Fix build warnings in RiaInterpolationTools.cpp

+++ ApplicationCode/Application/Tools/RiaInterpolationTools.cpp
@@ -33,7 +33,7 @@ bool almostEqual( double a, double b, double maxRelDiff = std::numeric_limits<do
     double fabsa = std::fabs( a );
     double fabsb = std::fabs( b );
     // Find the largest
-    float largest = ( fabsb > fabsa ) ? fabsb : fabsa;
+    double largest = ( fabsb > fabsa ) ? fabsb : fabsa;
     return ( diff <= largest * maxRelDiff );
 }
 
@@ -194,7 +194,7 @@ void RiaInterpolationTools::interpolateMissingValues( const std::vector<double>&
             {
                 // The last value is inf: extrapolate from two last data points
                 int prevSetIndex2 = findPreviousDataPoint( y, prevSetIndex - 1 );
-                index             = extrapolateRange( index, y.size(), prevSetIndex2, prevSetIndex, x, y );
+                index             = extrapolateRange( index, (int) y.size(), prevSetIndex2, prevSetIndex, x, y );
             }
             else if ( nextSetIndex != static_cast<int>( y.size() ) )
             {
